pyproject.toml
[tool.poetry]
name = "lecarb"
version = "0.1.0"
description = "Are We Ready For Learned Cardinality Estimation?"
authors = ["Weiyuan Wu <doomsplayer@gmail.com>", "Xiaoying Wang <wangxiaoying0369@gmail.com>", "Changbo Qu <quchangbo1990@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
contexttimer = "0.3"
pandas = "1.0.3"
numpy = "1.18.1"
docopt = "0.6"
psycopg2-binary = "2.8.4"
scikit-learn = "0.23.0"
torch = "1.5.0"
sqlparse = "0.3.1"
dask = "1.2.0"
toolz = "0.9.0"
cloudpickle = "1.2.1"
tables = "3.5.1"
spflow = "0.0.34"
bloom-filter = "1.3"
python-dotenv = "0.13.0"
jupyter = "1.0.0"
jupyterlab = "2.1.4"
seaborn = "0.11.0"
scipy = "1.4.1"
xgboost = "1.1.1"
ray = "^0.8.7"
Cython = "^0.29.21"
pomegranate = "^0.13.4"
pyodbc = "^4.0.30"
mysql-connector-python = "^8.0.21"
PyQt5 = "5.15.1"

[tool.poetry.dev-dependencies]
mypy = "0.770"
black = "19.10b0"
pylint = "2.4.4"
ipython = "7.13.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
